libs/py-file-helpers/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "py_file_helpers" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../coverage/py-file-helpers/html'
 --cov-report xml:'../coverage/py-file-helpers/coverage.xml'
 --html='../reports/py-file-helpers/unittests/html/index.html'
 --junitxml='../reports/py-file-helpers/unittests/junit.xml'
"""

[tool.poetry]
name = "py-file-helpers"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "py_file_helpers"

  [tool.poetry.dependencies]
  python = ">=3.10,<3.13"
  pyexiftool = "^0.5.6"
  pillow = "^12.0.0"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.3.1"
  flake8 = "7.1.1"
  pytest = "8.3.4"
  pytest-sugar = "1.0.0"
  pytest-cov = "6.0.0"
  pytest-html = "4.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
